Core dump when GetInt on a string
Was planning to structure my control logic in sum() and mean() using
exception returns. It seems like GetInt() core dumps when attempted on
a string, without generating a catchable exception. Because of this,
planning to refactor control logic to use IsNumber() and  IsUint64()
etc instead. This may actually streamline vector allocation too.